Remove AST attempt to resolve sample relationships
There are parallel versions of attempts to resolve liquid-liquid
relationships in the codebase in at least two places: ast and
microarch/scheduler/liquidhandling

While obviously a Bad Thing (and slated for revision) it's not
previously caused any semantic issues. However making MixerPrompt
into an N-N function instead of 1-1 unearthed an issue with the
ast version of this creating cycles in the instruction tree.

How exactly it did this is still not completely clear to me, however
since it's entirely unnecessary it's been stripped out in this change.

Test Plan: antha-tester runs cleanly against the elements generated from
the corresponding branch in elements-westeros, so nothing is broken.

Additionally some test bundles in ANTHA-2636 were run to ensure that
the original issue motivating this was solved