Remove PassThrough from LHInstruction
Prompts as originally implemented ended up a bit of a mess because
at that point we only allowed one component to be returned from
an instruction.

The consequence of this was that they were 1:1 from the user point
of view but N:N internally, the relationship being handled with a
map declaring which outputs related to which inputs. This map was
called PassThrough.

Now that instructions can have multiple returns it makes no sense to
keep this in place, so this diff removes the map and makes everything
simpler.

It's been tested by running against a few prompt-heavy workflows (which
went fine) - this is probably all that's required since we'd see runtime
errors if anything were amiss here. I'll also run antha-tester and paste
the results HERE