Convert plate types from inventory protobuf
antha-runner/aconv already contains converters from inventory plates
and plate types into the antha-lang versions, however the new
PlateType class is not supported.

Since it seems reasonable enough to have a method to extract a plate
type back out of a plate the conversion now works by creating a plate
from the inventory Plate Type then extracting the plate type from it.

I added a basic unit test which demonstrates that the extracted plate
type is fine for making another plate of the same type with, and equal
to the one which was used in the first place.