[AC-72] Limit element concurrency to 1.
Hopefully this should be temporary, but the wider problem is actually challenging.
For the time being, this throttle solves data races when you have a fork:
- three elements: A, B and C
- output of A is wired to B
- output of B is wired to C

with B and C running concurrently, they both turn out to access the exact same value (e.g. a Liquid), which it was never designed for.