[AC-73] Make exporting protobuf part of lab decommissioning
This solves part of the problem: if there are errors before we got to lab.RunElements (e.g. paramater parsing) then we wouldn't export the protobuf, which means we didn't get errors back to the user. Fixed.

However, this is only part of the problem. The other part is that we need to also export protobuf even if we don't get as far as execing the workflow.