[AC-73] Export errors to protobuf from the composer.
This is the other part of the fix: even when we don't get as far as starting the workflow, we need to export protobuf errors.
There is some slight complexity around making sure we don't accidentally create two response.pb files, but other than that, fairly simple.
Decided there's no point bothering with the "protobuf" build tag any more, so just destroy that for simplicity.