Fix/prompts to multiple args (#1039)
* Refactor prompt to take multiple args

The goal here is to ensure that prompting works for barriers
in a better way - currently aggregation is done after instruction
sorting, which means that some prompt barriers are split. The
underlying issue is that the prompt command was improperly specified
as 1-1 whereas it should be N-N. Modifying this means that now semantics
should be much better, liquids prompted on at the same time in a single
call should always remain together in time

* Fix Prompt aggregation

Prompt aggregation is a bit weird, and breaks if prompts are N-N
this fixes the issue

it remains to remove PassThrough

* Remove AST attempt to resolve sample relationships

There are parallel versions of attempts to resolve liquid-liquid
relationships in the codebase in at least two places: ast and
microarch/scheduler/liquidhandling

While obviously a Bad Thing (and slated for revision) it's not
previously caused any semantic issues. However making MixerPrompt
into an N-N function instead of 1-1 unearthed an issue with the
ast version of this creating cycles in the instruction tree.

How exactly it did this is still not completely clear to me, however
since it's entirely unnecessary it's been stripped out in this change.

Test Plan: antha-tester runs cleanly against the elements generated from
the corresponding branch in elements-westeros, so nothing is broken.

Additionally some test bundles in ANTHA-2636 were run to ensure that
the original issue motivating this was solved

(INSERT RESULTS HERE)

* Remove PassThrough from LHInstruction

Prompts as originally implemented ended up a bit of a mess because
at that point we only allowed one component to be returned from
an instruction.

The consequence of this was that they were 1:1 from the user point
of view but N:N internally, the relationship being handled with a
map declaring which outputs related to which inputs. This map was
called PassThrough.

Now that instructions can have multiple returns it makes no sense to
keep this in place, so this diff removes the map and makes everything
simpler.

It's been tested by running against a few prompt-heavy workflows (which
went fine) - this is probably all that's required since we'd see runtime
errors if anything were amiss here. I'll also run antha-tester and paste
the results HERE

* Prevent crash on JSON unmarshal

JSON unmarshalling, it turns out, calls NewMessageInstruction(nil)

previously we assumed the input instruction to be non-nil
this is now fixed.

Antha-tester now runs OK so this seems fine

Update: antha-tester succeeds for all bundles.

* Readability changes in codegen and execute

This is very much the tiniest tip of a very horrible iceberg
but it does very slightly improve the readability of the code
there

Tested by recompiling. Readability will be tested by passing it
by Haydn's eagle eyes

* Readability on 'cycle containing ... ' message

Just finishing the thought in the previous commit, this
message is hopefully a bit more informative

* Minor fixes to allow for testing

* Tests to highlight prompt behaviours

Unit tests to show the main cases of how prompt with multiple args
can work

- a) mix - prompt - mix on one liquid as a basic case
- b) adding prompts on two single liquids separately to show aggregation works
- c) case b) but replacing multiple single prompts with a prompt
     on both liquids to show this gives the same result
- d) a pathological case where aggregation of prompts causes an
     unexpected result
- e) the analogous case with the correct behaviour as a result of
     prompting on the two liquids simultaneously

* Added test to validate prompt input / output relation is correct

* Added test to validate prompt input / output relation is correct

Tested using antha-tester, passes successfully