Rework logging.
Logger now has a much more intuitive API and more obvious inheritance pattern when it comes to swapping out the underlying writers.
This has also enabled a slight widening of the API so that we can use the underlying writer directly in a safe way.
This means the composer now logs to disk everything that comes out of the workflow runtime, and the workflow runtime itself doesn't need to worry about logging to disk at all. So all the logs end up in one place which is vastly preferable.

unit tests pass, elements tests pass, linter passes.